requirements.txt
fastapi
uvicorn[standard]
scikit-learn 
scipy 
matplotlib 
numpy
pandas
nltk
SQLAlchemy
mysqlclient
mysql-connector-python